Add favorites profile support.
This is the first build of the server for client release 2.4.
